moving stat-card component into utilities directory, using min-h-screen instead of h-screen
